funny eror, lost track of my variables

+++ main.h
@@ -7,7 +7,7 @@ int _putchar(char c);
 int _puts(char *);
 int check(char str);
 int print_int(va_list args);
-char *convert(unsigned int num, int base);
+char *convert(long int num, int base);
 int _strlen(char *);
 int print_str(va_list);
 int print_char(va_list);
